requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
celery==5.3.4
redis==5.0.1
litellm==0.9.4
pydantic==2.4.2
python-dotenv==1.0.0
